Fixed d_term and dt computation
Cleaned up from quick fixes to make it run on hardware in the lab.
d_term was not calling the correct derivative, changed all to use
self.err_derivative. Also, this did not error previously because
dt was always set to 0.0, so Dan's non-~selfed~ derivative, set to
0.0, would be used to compute the d_term, though it technically
was never set or intended to be used. Now, self.time_prev should be
updating correctly, and dt should be computed correctly each time
the subscriber callback is triggered.